ExactRepository: move ownership of ExactApi out of repository.
This introduces a new ExactApi singleton, which now has the sole
ownership of the api. This makes it easier to split the single
large repository in multiple smaller ones in the future.

This also fixes a bug where some options could not be selected after
setting up the exact api.